[Update] Split math functions into aggregate and scalar ones
