+++ nujo/__init__.py
@@ -3,18 +3,8 @@ from nujo.math import (abs, max, mean, median, min, prod, sqrt, stddev, sum,
                        variance)
 
 __all__ = [
-    'Tensor'
+    'Tensor',
     'no_diff',
-    'abs',
-    'max',
-    'mean',
-    'median',
-    'min',
-    'prod',
-    'sqrt',
-    'stddev',
-    'sum',
-    'variance',
 ]
 
 __version__ = '0.1.0'
